+++ variables.tf
@@ -106,3 +106,30 @@ variable "file_share_backup_policy_retention" {
   type        = string
   default     = "30"
 }
+
+###############################
+# Azure Automation Account variables
+###############################
+variable "automation_account_sku" {
+  description = "Automation account Sku"
+  type        = string
+  default     = "Basic"
+}
+
+variable "custom_automation_account_name" {
+  description = "Automation account custom name"
+  type        = string
+  default     = ""
+}
+
+variable "law_resource_group_name" {
+  description = "Resource group of Log Analytics Workspace that will be connected with the automation account (default is the same RG that the one hosting the automation account)"
+  type        = string
+  default     = ""
+}
+
+variable "log_analytics_workspace_name" {
+  description = "Log Analytics Workspace that will be connected with the automation account"
+  type        = string
+  default     = ""
+}