+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.2.0 - Unreleased
+
+Added
+  * AZ-208: Add automation account module
+
 # v2.1.0 - 2020-01-31
 
 Added